+++ src/main/java/io/github/astrapi69/file/csv/CsvFileExtensions.java
@@ -40,16 +40,18 @@ import java.util.Set;
 import java.util.StringTokenizer;
 
 import io.github.astrapi69.file.write.StoreFileExtensions;
-import io.github.astrapi69.file.write.WriteFileExtensions;
 import io.github.astrapi69.io.StreamExtensions;
 import io.github.astrapi69.string.StringExtensions;
 
 /**
  * Utility class for the use of cvs-files.
  *
+ * @deprecated Use instead the class with the same name from module 'cvs-worker'. Note: will be
+ *             removed with next minor version
  * @version 1.0
  * @author Asterios Raptis
  */
+@Deprecated
 public final class CsvFileExtensions
 {
 
@@ -626,7 +628,7 @@ public final class CsvFileExtensions
 	public static void writeLinesToFile(final Collection<String> collection, final File output,
 		final String encoding) throws IOException
 	{
-		WriteFileExtensions.writeLinesToFile(collection, output, encoding);
+		StoreFileExtensions.toFile(output, collection, encoding);
 	}
 
 }