+++ src/main/java/io/github/astrapi69/file/delete/DeleteFileExtensions.java
@@ -49,6 +49,10 @@ import io.github.astrapi69.io.file.filter.PrefixFileFilter;
  */
 public final class DeleteFileExtensions
 {
+
+	/**
+	 * Private constructor to prevent instantiation
+	 */
 	private DeleteFileExtensions()
 	{
 	}